# docs/api.yml
openapi: 3.0.0
info:
  title: Markdown Note System API
  version: 1.0.0
  description: API for Markdown Note System
servers:
  - url: http://localhost:8080/api/v1
paths:
  /auth/login:
    post:
      summary: User Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /documents:
    get:
      summary: List documents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of documents
    post:
      summary: Create new document
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '201':
          description: Created
  /documents/{id}:
    get:
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
    put:
      summary: Update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Document'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted

components:
  schemas:
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        categoryId:
          type: integer
        tags:
          type: array
          items:
            type: string